+++ CHANGELOG.md
@@ -16,6 +16,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
 - [@Rsullivan00] - Prevent `command` and `command_call` nodes from being turned into ternaries.
 - [@kddeisz] - Better handling of the `alias` node with and without comments.
 - [@kddeisz] - Better handling of the `BEGIN` and `END` nodes with and without comments.
+- [@kddeisz] - Much better handling of heredocs where now there is a consistent `heredoc` node instead of multiple.
 
 ## [0.20.1] - 2020-09-04