+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
 
 ## [Unreleased]
 
+## [0.21.0] - 2020-12-02
+
 ### Changed
 
 - [@kddeisz], [@ryan-hunter-pc] - Explicitly handle `break` and `next` keyword parentheses.
@@ -860,7 +862,8 @@ would previously result in `array[]`, but now prints properly.
 
 - Initial release 🎉
 
-[unreleased]: https://github.com/prettier/plugin-ruby/compare/v0.20.1...HEAD
+[unreleased]: https://github.com/prettier/plugin-ruby/compare/v0.21.0...HEAD
+[0.21.0]: https://github.com/prettier/plugin-ruby/compare/v0.20.1...v0.21.0
 [0.20.1]: https://github.com/prettier/plugin-ruby/compare/v0.20.0...v0.20.1
 [0.20.0]: https://github.com/prettier/plugin-ruby/compare/v0.19.1...v0.20.0
 [0.19.1]: https://github.com/prettier/plugin-ruby/compare/v0.19.0...v0.19.1